UPSTREAM: block: silently forbid sending any ioctl to a partition
cherry picked from:
 commit 0478fe68685a ("block: silently forbid sending any ioctl to a partition")

After the first few months, the message has not led to many bug reports.
It's been almost five years now, and in practice the main source of
it seems to be MTIOCGET that someone is using to detect tape devices.
While we could whitelist it just like CDROM_GET_CAPABILITY, this patch
just removes the message altogether.

The patch also removes the "safe but not very useful" ioctl whitelist,
as suggested by Christoph.  I doubt anything is using most of those
ioctls _in general_, let alone on a partition.

Bug: 73949641
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Change-Id: Iece98c5f31c5b5c03afe3587fab74a0233a8e1e4
Signed-off-by: TogoFire <togofire@mailfence.com>