cpu-boost: Bind to LITTLE cpus and reduce RT prio
To prevent the case of a possible RT priority inversion with more
important RT tasks like mdss (RIP), reduce the priority to 2.  This will
make sure cpu-boost can take priority over most work items while not
causing a priority inversion leading to thread interruption.

Additionally we bind the cpu-boost worker thread to CPUs 0-3 to decrease
the likeliness of it interfering with the app threads.  For that
purpose, we do a small refactor of the kthread init process.

While at it, add logging for the case where sched_setscheduler fails, as
it can fail due to security checks.

Signed-off-by: Alex Naidis <alex.naidis@paranoidandroid.co>
Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>
Signed-off-by: iusmac <iusico.maxim@libero.it>