Revert "kernel: Only expose su when daemon is running"
This patch is no longer necessary because we no longer ship
su add-ons, which is this patch initially designed for. Now
it causes another issue which breaks custom root solution
such as Magisk, as Magisk switches worker tmpfs dir to RO
instead of RW for safety reasons and happens to satisfy
MS_RDONLY check for su file, resulting in su file totally
inaccessible.

This reverts commit 64fc568ba241d0fee25b8bf64c05b5f29312a24d.
This reverts commit 6c4061c21759391328576064d72da6d89604b947.