Revert "BACKPORT: cpufreq: schedutil: Avoid using invalid next_freq"
"next_freq" is never set to UINT_MAX. While this change was made in
upstream with commit ecd288429126 ("cpufreq: schedutil: Don't set
next_freq to UINT_MAX"), it also discarded commit 97739501f207
("cpufreq: schedutil: Avoid using invalid next_freq").

So, the current patch was mistakenly added to the tree that just
breaks the intended functionality and is simply proving to be
useless. Revert it.

This reverts commit e982df9aa3dd5007dc70cc2e37b01c4dc2e67c32.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Change-Id: Ideae3b711cc9c2b3b6086751b199ed74261609bc
Signed-off-by: TogoFire <togofire@mailfence.com>